+++ receiver/hostmetricsreceiver/internal/scraper/cpuscraper/internal/metadata/generated_metrics_v2.go
@@ -25,7 +25,7 @@ func DefaultMetricsSettings() MetricsSettings {
 			Enabled: true,
 		},
 		SystemCPUUtilization: MetricSettings{
-			Enabled: true,
+			Enabled: false,
 		},
 	}
 }